Добавлена функциональность «Лента событий»

+++ src/main/java/ru/yandex/practicum/filmorate/model/event/Event.java
@@ -11,13 +11,13 @@ public class Event {
     @Positive
     private long timestamp;
     @Positive
-    private Long userId;
+    private Integer userId;
     @NotNull
     private EventType eventType;
     @NotNull
     private Operation operation;
     @Positive
-    private Long eventId;
+    private Integer eventId;
     @Positive
-    private Long entityId;
+    private Integer entityId;
 }
